+++ hvcc/__init__.py
@@ -19,9 +19,9 @@ import json
 import os
 import re
 import time
+import sys
 
 from hvcc.interpreters.pd2hv import pd2hv
-from hvcc.interpreters.max2hv import max2hv
 from hvcc.core.hv2ir import hv2ir
 from hvcc.generators.ir2c import ir2c
 from hvcc.generators.ir2c import ir2c_perf
@@ -162,21 +162,14 @@ def compile_dataflow(in_path, out_dir, patch_name=None, patch_meta_file=None,
     patch_name = patch_name or "heavy"
     generators = generators or {"c"}
 
-    if in_path.endswith((".pd", ".maxpat")):
+    if in_path.endswith((".pd")):
         if verbose:
             print("--> Generating C")
-        if in_path.endswith(".pd"):
-            results["pd2hv"] = pd2hv.pd2hv.compile(
-                pd_path=in_path,
-                hv_dir=os.path.join(out_dir, "hv"),
-                search_paths=search_paths,
-                verbose=verbose)
-        elif in_path.endswith(".maxpat"):
-            results["max2hv"] = max2hv.max2hv.compile(
-                max_path=in_path,
-                hv_dir=os.path.join(out_dir, "hv"),
-                search_paths=search_paths,
-                verbose=verbose)
+        results["pd2hv"] = pd2hv.pd2hv.compile(
+            pd_path=in_path,
+            hv_dir=os.path.join(out_dir, "hv"),
+            search_paths=search_paths,
+            verbose=verbose)
 
         # check for errors
         if list(results.values())[0]["notifs"].get("has_error", False):
@@ -246,120 +239,62 @@ def compile_dataflow(in_path, out_dir, patch_name=None, patch_meta_file=None,
     num_input_channels = hvir["signal"]["numInputBuffers"]
     num_output_channels = hvir["signal"]["numOutputBuffers"]
 
+    gen_args = {
+        'c_src_dir': c_src_dir,
+        'out_dir': out_dir,
+        'patch_name': patch_name,
+        'patch_meta': patch_meta,
+        'num_input_channels': num_input_channels,
+        'num_output_channels': num_output_channels,
+        'externs': externs,
+        'copyright': copyright,
+        'verbose': verbose
+    }
+
     if "bela" in generators:
         if verbose:
             print("--> Generating Bela plugin")
-        results["c2bela"] = c2bela.c2bela.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "bela"),
-            patch_name=patch_name,
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            verbose=verbose)
+        results["c2bela"] = c2bela.c2bela.compile(**gen_args)
 
     if "fabric" in generators:
         if verbose:
             print("--> Generating Fabric plugin")
-        results["c2fabric"] = c2fabric.c2fabric.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "fabric"),
-            patch_name=patch_name,
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            copyright=copyright,
-            verbose=verbose)
+        results["c2fabric"] = c2fabric.c2fabric.compile(**gen_args)
 
     if "js" in generators:
         if verbose:
             print("--> Generating Javascript")
-        results["c2js"] = c2js.c2js.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "js"),
-            patch_name=patch_name,
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            copyright=copyright,
-            verbose=verbose)
+        results["c2js"] = c2js.c2js.compile(**gen_args)
 
     if "daisy" in generators:
         if verbose:
             print("--> Generating Daisy module")
-        results["c2daisy"] = c2daisy.c2daisy.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "daisy"),
-            patch_name=patch_name,
-            patch_meta=patch_meta,
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            copyright=copyright,
-            verbose=verbose)
+        results["c2daisy"] = c2daisy.c2daisy.compile(**gen_args)
 
     if "dpf" in generators:
         if verbose:
             print("--> Generating DPF plugin")
-        results["c2dpf"] = c2dpf.c2dpf.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "plugin"),
-            patch_name=patch_name,
-            patch_meta=patch_meta,
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            copyright=copyright,
-            verbose=verbose)
+        results["c2dpf"] = c2dpf.c2dpf.compile(**gen_args)
 
     if "owl" in generators:
         if verbose:
             print("--> Generating OWL plugin")
-        results["c2owl"] = c2owl.c2owl.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "Source"),
-            patch_name=patch_name,
-            copyright=copyright,
-            verbose=verbose)
+        results["c2owl"] = c2owl.c2owl.compile(**gen_args)
 
     if "pdext" in generators:
         if verbose:
             print("--> Generating Pd external")
-        results["c2pdext"] = c2pdext.c2pdext.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "pdext"),
-            patch_name=patch_name,
-            ext_name=patch_name + "~",
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            copyright=copyright,
-            verbose=verbose)
+        results["c2pdext"] = c2pdext.c2pdext.compile(**gen_args)
 
     if "unity" in generators:
         if verbose:
             print("--> Generating Unity plugin")
-        results["c2unity"] = c2unity.c2unity.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "unity"),
-            patch_name=patch_name,
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            copyright=copyright,
-            verbose=verbose)
+        results["c2unity"] = c2unity.c2unity.compile(**gen_args)
 
     if "wwise" in generators:
         if verbose:
             print("--> Generating Wwise plugin")
-        results["c2wwise"] = c2wwise.c2wwise.compile(
-            c_src_dir=c_src_dir,
-            out_dir=os.path.join(out_dir, "wwise"),
-            patch_name=patch_name,
-            num_input_channels=num_input_channels,
-            num_output_channels=num_output_channels,
-            externs=externs,
-            verbose=verbose)
+        results["c2wwise"] = c2wwise.c2wwise.compile(**gen_args)
 
     return results
 
@@ -422,15 +357,18 @@ def main():
         verbose=args.verbose,
         copyright=args.copyright)
 
+    errorCount = 0
     for r in list(results.values()):
         # print any errors
         if r["notifs"].get("has_error", False):
             for i, error in enumerate(r["notifs"].get("errors", [])):
+                errorCount += 1
                 print("{4:3d}) {2}Error{3} {0}: {1}".format(
                     r["stage"], error["message"], Colours.red, Colours.end, i + 1))
 
             # only print exception if no errors are indicated
             if len(r["notifs"].get("errors", [])) == 0 and r["notifs"].get("exception", None) is not None:
+                errorCount += 1
                 print("{2}Error{3} {0} exception: {1}".format(
                     r["stage"], r["notifs"]["exception"], Colours.red, Colours.end))
 
@@ -454,7 +392,9 @@ def main():
 
     if args.verbose:
         print("Total compile time: {0:.2f}ms".format(1000 * (time.time() - tick)))
+    return errorCount != 0
 
 
 if __name__ == "__main__":
-    main()
+    ret = main()
+    sys.exit(ret)