+++ hvcc/generators/c2dpf/c2dpf.py
@@ -29,6 +29,7 @@ class c2dpf:
 
         tick = time.time()
 
+        out_dir = os.path.join(out_dir, "plugin")
         receiver_list = externs['parameters']['in']
 
         if patch_meta:
@@ -36,7 +37,9 @@ class c2dpf:
             dpf_meta = patch_meta.get("dpf", {})
         else:
             dpf_meta = {}
+
         dpf_project = dpf_meta.get('project')
+        dpf_path = dpf_meta.get('dpf_path', '')
 
         copyright_c = copyright_manager.get_copyright_for_c(copyright)
         # copyright_plist = copyright or u"Copyright {0} Enzien Audio, Ltd." \
@@ -109,14 +112,16 @@ class c2dpf:
             with open(os.path.join(source_dir, "Makefile"), "w") as f:
                 f.write(env.get_template("Makefile").render(
                     name=patch_name,
-                    meta=dpf_meta))
+                    meta=dpf_meta,
+                    dpf_path=dpf_path))
 
             # project makefile
             if dpf_project:
                 with open(os.path.join(source_dir, "../../Makefile"), "w") as f:
                     f.write(env.get_template("Makefile.project").render(
                         name=patch_name,
-                        meta=dpf_meta))
+                        meta=dpf_meta,
+                        dpf_path=dpf_path))
 
             buildjson.generate_json(
                 out_dir,