+++ hvcc/generators/ir2c/SignalConvolution.py
@@ -47,7 +47,7 @@ class SignalConvolution(HeavyObject):
         ]
 
     @classmethod
-    def get_C_process(clazz, obj_type, process_dict, args):
+    def get_C_process(clazz, process_dict, obj_type, obj_id, args):
         return [
             "__hv_conv_f(&sConv_{0}, VIf({1}), VOf({2}));".format(
                 process_dict["id"],