+++ hvcc/generators/ir2c/templates/Heavy_NAME.cpp
@@ -2,8 +2,9 @@
 
 #include "Heavy_{{name}}.hpp"
 
-#define Context(_c) reinterpret_cast<Heavy_{{name}} *>(_c)
+#include <new>
 
+#define Context(_c) static_cast<Heavy_{{name}} *>(_c)
 
 
 /*
@@ -12,12 +13,31 @@
 
 extern "C" {
   HV_EXPORT HeavyContextInterface *hv_{{name}}_new(double sampleRate) {
-    return new Heavy_{{name}}(sampleRate);
+    // allocate aligned memory
+    void *ptr = hv_malloc(sizeof(Heavy_{{name}}));
+    // ensure non-null
+    if (!ptr) return nullptr;
+    // call constructor
+    new(ptr) Heavy_{{name}}(sampleRate);
+    return Context(ptr);
   }
 
   HV_EXPORT HeavyContextInterface *hv_{{name}}_new_with_options(double sampleRate,
       int poolKb, int inQueueKb, int outQueueKb) {
-    return new Heavy_{{name}}(sampleRate, poolKb, inQueueKb, outQueueKb);
+    // allocate aligned memory
+    void *ptr = hv_malloc(sizeof(Heavy_{{name}}));
+    // ensure non-null
+    if (!ptr) return nullptr;
+    // call constructor
+    new(ptr) Heavy_{{name}}(sampleRate, poolKb, inQueueKb, outQueueKb);
+    return Context(ptr);
+  }
+
+  HV_EXPORT void hv_{{name}}_free(HeavyContextInterface *instance) {
+    // call destructor
+    Context(instance)->~Heavy_{{name}}();
+    // free memory
+    hv_free(instance);
   }
 } // extern "C"